api/requirements.txt
fastapi
uvicorn[standard]
Pillow
opencv-python
ultralytics
transformers
torch